git-svn-id: https://scm.projects.hlrs.de/authscm/jenkinsls1/svn/ls1/MarDyn/branches/mheinenNEMD/mainbranch20170119@5272 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/utils/xmlfile.h
@@ -594,8 +594,8 @@ public:
 	
 	/// \brief number of registered queries
 	/// return the number of active queries
-	/// \param unsigned int	number of queries
-	unsigned int numqueries() const
+	/// \return number of active queries
+	size_t numqueries() const
 		{ return m_queries.size(); }
 
 #ifdef ENABLE_MPI
